some fixes

+++ ragemp-udp/lib/SLikeNet/DependentExtensions/cat/src/crypt/hash/Skein.cpp
@@ -39,7 +39,7 @@ Skein::~Skein()
 
 void Skein::GenerateInitialState(int bits)
 {
-    u64 w[MAX_WORDS] = { getLE64(SCHEMA_VER), getLE64(static_cast<unsigned>(bits)), 0 };
+    u64 w[MAX_WORDS] = { getLE64(SCHEMA_VER), getLE64(bits), 0 };
 
     CAT_OBJCLR(State);
